vidhi-backend/requirements.txt
# ═══════════════════════════════════════════════════════════
# VIDHI Backend — Python Requirements
# Python 3.11+ recommended (NOT 3.14 — ChromaDB incompatible)
# ═══════════════════════════════════════════════════════════

# ──────────────────────────────
# API Framework
# ──────────────────────────────
fastapi>=0.110.0,<0.116.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.0.0,<3.0.0
python-dotenv==1.0.0

# ──────────────────────────────
# AWS SDK
# ──────────────────────────────
boto3>=1.34.0
botocore>=1.34.0

# ──────────────────────────────
# LangChain + RAG Stack
# ──────────────────────────────
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-aws>=0.1.0,<0.2.0
langchain-text-splitters>=0.2.0,<0.3.0

# ──────────────────────────────
# Vector Store
# ──────────────────────────────
# NOTE: ChromaDB is NOT compatible with Python 3.14+
chromadb==0.4.22
langchain-chroma>=0.1.0,<0.2.0

# ──────────────────────────────
# Document Processing
# ──────────────────────────────
beautifulsoup4==4.12.3
lxml==5.1.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# ──────────────────────────────
# Web Scraping (for data pipeline)
# ──────────────────────────────
selenium==4.17.2
requests>=2.31.0
httpx==0.26.0

# ──────────────────────────────
# Data Pipeline
# ──────────────────────────────
datasets>=2.0.0
huggingface-hub>=0.20.0

# ──────────────────────────────
# Utilities
# ──────────────────────────────
python-dateutil==2.8.2
pytz==2024.1

# ──────────────────────────────
# AWS Lambda Deployment (optional)
# ──────────────────────────────
mangum==0.17.0

# ──────────────────────────────
# Logging
# ──────────────────────────────
python-json-logger==2.0.7

# ──────────────────────────────
# Testing
# ──────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.3

# ──────────────────────────────
# Dev Tools
# ──────────────────────────────
black==24.1.1
flake8==7.0.0
